+++ KickLib/Api/Interfaces/ICategories.cs
@@ -3,7 +3,7 @@
 namespace KickLib.Api.Interfaces;
 
 /// <summary>
-/// Category operations
+///     Category operations.
 /// </summary>
 public interface ICategories
 {
@@ -11,10 +11,10 @@ public interface ICategories
     ///     Gets categories or search for category by name.
     ///     Returns up to 100 results at a time; use the <paramref name="page"/> parameter to get more results.
     /// </summary>
-    Task<Result<ICollection<CategoryResponse>>> GetCategoriesAsync(string searchKeyword, int? page, string? accessToken, CancellationToken cancellationToken);
+    Task<Result<ICollection<CategoryResponse>>> GetCategoriesAsync(string searchKeyword, int? page, string? accessToken = null, CancellationToken cancellationToken = default);
 
     /// <summary>
     ///     Gets specific category by ID.
     /// </summary>
-    Task<Result<CategoryResponse>> GetCategoryAsync(int id, string? accessToken, CancellationToken cancellationToken);
+    Task<Result<CategoryResponse>> GetCategoryAsync(int id, string? accessToken = null, CancellationToken cancellationToken = default);
 }