+++ KickLib/Api/Interfaces/IChat.cs
@@ -12,23 +12,23 @@ public interface IChat
     ///     Post a chat message to a channel as a user.
     ///     When sending as a user, the broadcaster_user_id is required.
     /// </summary>
-    Task<Result<SendChatMessageResponse>> SendMessageAsUserAsync(int broadcasterId, string message, string? accessToken, CancellationToken cancellationToken);
+    Task<Result<SendChatMessageResponse>> SendMessageAsUserAsync(int broadcasterId, string message, string? accessToken = null, CancellationToken cancellationToken = default);
 
     /// <summary>
     ///     Post a chat message to a channel as a bot.
     ///     As a bot, the message will always be sent to the channel attached to your token.
     /// </summary>
-    Task<Result<SendChatMessageResponse>> SendMessageAsBotAsync(string message, string? accessToken, CancellationToken cancellationToken);
+    Task<Result<SendChatMessageResponse>> SendMessageAsBotAsync(string message, string? accessToken = null, CancellationToken cancellationToken = default);
 
     /// <summary>
     ///     Post a chat message as reply to a channel as a user.
     ///     When sending as a user, the broadcaster_user_id is required.
     /// </summary>
-    Task<Result<SendChatMessageResponse>> ReplyToMessageAsUserAsync(int broadcasterId, string message, string messageId, string? accessToken, CancellationToken cancellationToken);
+    Task<Result<SendChatMessageResponse>> ReplyToMessageAsUserAsync(int broadcasterId, string message, string messageId, string? accessToken = null, CancellationToken cancellationToken = default);
 
     /// <summary>
     ///     Post a chat message as reply to a channel as a bot.
     ///     As a bot, the message will always be sent to the channel attached to your token.
     /// </summary>
-    Task<Result<SendChatMessageResponse>> ReplyToMessageAsBotAsync(string message, string messageId, string? accessToken, CancellationToken cancellationToken);
+    Task<Result<SendChatMessageResponse>> ReplyToMessageAsBotAsync(string message, string messageId, string? accessToken = null, CancellationToken cancellationToken = default);
 }