+++ KickLib/Core/KickLibHttpClientFactory.cs
@@ -1,0 +1,9 @@
+namespace KickLib.Core;
+
+internal class KickLibHttpClientFactory : IHttpClientFactory
+{
+    public HttpClient CreateClient(string? name = null)
+    {
+        return new HttpClient();
+    }
+}